[assembler-arm64] Disable an assertion on uncompressed CHERI.
When we do something like a Math.ceil(0), we will call into the Torque
runtime function which takes a type of JSAny. In there, we will
potentially cast the capability that JSAny is into a Smi and generate an
extended add instruction. In the backend, we use CnCSP() to do so. For
now, just disable the IsZero() check in CnCSP() to move forward with the
port.