[cheri] Change digit_t and signed_digit_t types.
On CHERI architectures intptr_t and uintpr_t are capability preserving
types, and therefore store a capability type twice the native integer
pointer size of the baseline architecture. Defining digit_t and
signed_digit_t using u/intptr_t is therefore clearly not correct; and
furthermore is a misue of the type on conventional architectures.
Therefore, these types are defined as fixed width integer types
depending on the baseline architecture.