[cheri] Replace cast of kve_start and kve_end.
On CHERI architecturs the reinterpret_cast from uint64_t to uintptr_t
fails as uintptr_t is a capability poreserving type and therefore twice
the native integer pointer size of the baseline architecture. Top mimise
the changes and impact on other platforms the cast has been reverted to
a static_cast as was present in earlier versions of V8.