[cheri] Correct usage of size_t/uintptr_t types.
On CHERI architectures size_t and uintptr_t are not the same size. In a
number of places the V8 codes is fairly lax with these types assuming
them to be the same size. In this case using size_t for the page_base,
invalidates the capability and results in a warning.