[cheri] Add new CountTrailingZeros/Population.
The v8 bit vector is implemented with a an inline value and a pointer.
On CHERI the inline value is now twice the width of the native integer
pointer size of the baseline argument. On 64-bit architectures the bit
vector is 128-bits wide. These additions support couting the trailing
zeros and populations of 128-bits values.

Mirror changes in CountLeadingZeros.