[cheri] User non-capability types for mem chunk.
A number of fields used capability types to store integer values and
differences between pointers (such as the high watermark). To provent
the addition of excessive padding into the offset calculations these
types have been changed to more suitable, non-capability types.

The static_asserts on memory chunks heasder sizes and and so on all
appear to be correct at this time.