[cheri] Update static_assert of AnyCType size.
On 64-bit CHERI architectures, AnyCType is 16-bytes (128-bits). AnyCType
is used by the arm64 simulator and the assumption of 64-bits is clearly
present there as parameters are stored in X registers or pushed to the
stack. However, the original assert states that the
EffectControlLinearizer assumes AnyCType is 64-bits, this is currently a
bit unclear. And may cause problems in future due to to the increase
size on CHERI architectures.