[cheri] Hash u/intptr_t and pointer values.
CHERI C/C++ Programming Guide: Computing hash values

The compiler will also warn when operators such as modulus or shifts are
used on uintptr_t. This usually indicates that the pointer is being used
as the input to a hash function or similar computations. In this case,
the programmer should not be using uintptr_t but instead cast the
pointer to ptraddr_t and perform the arithmetic on this type instead.