[cheri] Adapt static_assert fo capabilities.
Code Stub Assembler uses IntPtr for integer values, hence the check as
to whether the size matches that of size_t. Longer term this must be
change on CHERI as using a capability carrying type for an integer value
is wasteful and confusing as it makes it unclear where a capability
speciifc load/store/add should be generated.