[cheri] Disable setting v8_falgs read-only.
On CheriBSD the class to base::OS::SetDataReadOnly results in a
protection fault EPROT; the exact cause of this fault isn't understood.
Although CHERI offers the possbility of monotonically decreasing the
permissions of v8_flags, removing the possibility of setting them
writtable, this requires some redesign of the flags. As this is
currently not the priority this is currently left for future work.