[heap] Change how we align capabilities for CHERI.
Prior to this commit, we would align all the capabilities at the
"middle" and "top" levels of the allocators. However, this would result
in attempting to access maps at the wrong (unaligned) addresses later on
in the runtime. This commit makes it so that we align all the
allocations that need to be aligned at the "lower" levels of the
allocators, ensuring that the addresses we access later on are the
aligned (correct) ones.

This is still a work in progress and might change, as there are still
issues relating to alignment of allocations and loss of tags (usually by
getting the wrong address).