[serializer] Output per 8 bytes rather than capsize on CHERI.
When serializing the read-only heap, we can't always do
capability-sized output because some objects have content that is:
  (1) < 16 bytes
  (2) Not 16-byte aligned

This commit hardcodes the path to work on an 8-byte basis rather than
capsize-byte basis for uncompressed CHERI builds.

This might be the wrong approach to it, but it gets mksnapshot running
successfully. There's an issue when deserializing the snapshot, but it
is currently unclear if it's due to this commit or the deserializer
simply needs further modification for CHERI.