[cheri] Correctly align kInstructionStartOffset.
The Code data field kInstructionStartOffset is a pointer value,
therefore on CHERI this is a capability value that must be correctly
aligned. As Code is a heap object it is aligned to Tagged_t. On a
compressed pointer build padding is added so that the field maybe
correctly algiined. New methods have been added to read and write a
field with specific alignment requirements.