[cheri] Remove extraneous padding in offsets.
The marking CellType has been changed to a non-capability type, that is
size_t. As a result this manually padding in the offset calculations is
not required. Its presence resulted in incorrectly indexing the marking
map in the object, and as a result the object was shown is being marked
when it wasn't.