[cheri] Fixes to indexing of table.
Ensure that all indexing accounts for the `fat` entries with the pointer
and the tag being stored seperately.

Remove padding in the Payload struct as this prevents the atomic compare
and swap operations on the payload succeeding.

On CHERI architectures invalidating the pointer through tagging results
in an invalidated capability (that can't be restored, and there is no
obvious way to rederive the capability). Storing the pointer and the tag
seperated exposes the capability to type confusion attacks.  owever, it
should be noted the SW sandbox design is in conflict with HW protections
offered by CHERI and would likely be signifciantly redesigned in a
comprehensive porting activity.