[cheri] Ensure embedded code blobs are the same.
On CHERI the checks on tear down fail because the current, sticky, and
embedeed code blob are not the same. This is because the embedded code
blob is remapped into the code range. This appears to be a latent bug in
v8. The remapping has been updated to ensure that the current and sticky
embedded code blob match the current embedded code blob. Further
analysis of the conditions that casue this problem is required to
confirm that this is a latent bug (if it is a bug in v8 its unclear why
it doesn't manifest on other architectures).