[cheri] Disable checksum of embedded blob.
The checksum of the embedded blob attempts to treat code as data. On
CHERI accessing a sentry value as data produces a protection fault. To
resolve this more gracefully the instructions could be stored as code
and data, or a capability could be rederived through a call to the
runtime linker. As neither of these options is particularly elegant the
checksum is simply disabled for the time being.