[cheri] Treat embedded_blob_code as data.
The embedded_blob_code is iterated using the sizes of each function. If
this written to a text segment the code attempts to increment from the
initial sentry value invalidating the capability tag. Instead the
embedded_blob_code is treated as data and sentry values are created for
each function.

This approach prevents embedded builtins working whithout the JIT (that
is, executable memory) being enabled. Further work is needed to find a
solution that works on CHERI in both cases.

Set the maximum permissions for pages containing the embedded_blob_code
to read, write, execute.