[codegen] Re-derive sentries when we modify them.
Similar to cfa1aba25f5e47009064f91e1408f1b4ad72dd74, sometimes v8's
codegen modifies sentries. In order to preserve them, we need to
generate an instruction sequence as follows:

  adr c1, 0
  orr x0, x0, #1
  scvalue c1, c1, x0
  seal c0, c1, rb

However, we also need to check if we are actually operating on a sentry,
because we might be working with JITted code that simply has an RX
capability. We don't produce sentries for those yet, so we either derive
from the PCC, or we don't derive anything.

This commit further adds various Morello instructions such as gclen,
scbndse, gcseal, seal and build and re-formats some existing parts of
the code. It also adds a new abort reason for the codegen to use in case
of unexpected sealed capabilities and makes use of it in the Mov
immediate instruction, which should help debugging.