[compiler] Add kArm64StrPairCapability.
This commit adds a new IR instruction which represents a store of a pair
of capabilities. This was added in order to fix the bug where any StrPair
that needed to perform a sub instruction before actually storing to the
base register would be performed on X registers.

As a result, this commit also adds helpers to generate a capability
register base for a StrPairCapability and fixes a bug in the codegen.

Furthermore, remove a comment which is not applicable for purecap
builds in StoreReturnAddressAndCall.