[codegen] Allow mixing X/C registers in register lists.
Justification: Unlike W and X registers, X and C registers may be
treated interchangably depending on the processor mode on Morello, so it
doesn't make much sense to require them to be exact. Even though we did
not hit this case in particular, it creates a problem in one of the
tests that tries to include X registers in a scratch list.