[codegen] Fix CopyCapabilities, CopySlots, re-add CopyDoubleWords.
When we Subs two pointers and store the result into a pointer, we have
two codegen possibilities on Morello:
 - Generate a subs <xd>, <cn>, <cm>
 - Generate a sequence of gcvalue/scvalue instructions with a subs

In the former case, we will invalidate the capability's tag, which we
don't want since the register is later used as if it still has a valid
capability. In the latter case, we will probably go so far out of
representable bounds of the capability that we'll invalidate the tag as
well, creating the same problem as in the first case.

Since our goal is to simply compare the two pointers for an in-bounds
check, use a temporary X register to perform the comparison and don't
touch the registers that actually hold the capabilities.

This commit also adds a way to support X-register sized stack slot
addresses and ports a subset of the relevant tests.