[codegen] Rework the constant pool.
This commit adds the ability to store capabilities into the constant
pool by separating out pointers from integers. Instead of storing
embedded objects into 64-bit integer slots, we now store their index
into the pointer slot so that later on when we resolve the relocation,
we can actually fit the capability into it and perform a pc-relative
load to fetch it into a register.

This commit also fixes a number of codegen issues with the constant pool
that were present on purecap.