[codegen] Initial support for ExternalReference calls.
On CHERI, when calling external references we can't just
 Mov(temp, target);

and expect to call it successfully, as it won't have a tag. Further,
deriving from the PCC is not viable because JITted code might have a PCC
that is different from the one that holds the external reference.
Instead, we have to perform a pc-relative load of a capability and
generate a constant table to do it from. However, we can't do this from
the Call macro assembler method because it makes all the builtins that
make use of Abort() isolate-dependent. Thus, we special-case the Printf
codegen to perform such a call manually and patch the external reference
call with a brk #0x0 instruction in order to detect failures at runtime
early.

Furthermore, implement the varargs calling convention for Morello in
order to be able to successfully call printf(). We leverage the existing
argument shuffling algorithm in v8, even though it's not optimal for
what we are doing simply because this is largely a debugging tool and
performance doesn't really matter.

This commit also ports the relevant printf tests to CHERI.