[codegen] Fix external calls.
When generating a call to an external, there is an assembler option
which helps us identify if the code we generate will survive process
shutdown (i.e. being serialized). In the case it won't, no relocation
will be needed and V8 would attempt to generate a sequence of mov and
movk instructions in order to set up an address that it directly moves
into a register. Instead, use a pcrel load-literal instruction and
generate a constant pool which will store our capability in order to
perform the call.

We need to perform the relocation check outside of the
MacroAssembler::Mov() because moving a simple immediate into a register
isn't the same as moving a pointer into it -- we need to distinguish
between these two things, so instead of complicating Mov even further on
CHERI, simply do it in the place it's necessary.