[codegen] Move tag preserving moves into Mov().
The previous idea to keep it away from Mov doesn't really work because
there are a number of CSA functions that rely on correctly preserving
tags of pointers. Instead of guessing what is and isn't a pointer,
check if the immediate is tagged. If so, generate the constant pool and
perform a load-literal. If it isn't, perform a regular sequence of mov
and movk for moving an immediate.