[compiler] Handle pointer constants in CSA.
This is a rather large change. V8 tries very hard to reduce an IntPtrT
(and any intptr_t constant) into a 64-bit (or 32-bit) word in CSA. This
doesn't work for CHERI because there are cases when we really want to
load a pointer into a register.

Instead of trying to fix all type usage throughout V8 and CSA, we take
the pragmatic approach. We make use of kCapability64 (and define new
enum entries in appropriate places), kCapability64Constant and
kRelocatableCapability64. Then, when working with intptr_t constants, we
check if the value is actually a pointer by checking its tag. If it is,
we generate the capability versions of constants, and if it's not we
reduce them in-place to the a machine word (in practice on Morello, this
is always a 64-bit word).

In the backend of the codegen, we now need to call MacroAssembler::Ldr()
with a pointer literal in order to generate a constant pool to store the
capability into and load from during code execution. However, this
requires a relocation. Since none of the existing relocation types
are fit for purpose here, we create a new one: CAPABILITY_CONSTANT. This
relocation's only purpose is to emit capability constants into a
constant pool for use in a pcrel load-literal instruction and nothing
else should attempt to use them. They are currently only generated in
CSA.