[CSA] Initial pointer alignment on CHERI.
The CSA codegen isn't particularly well-tested and has known issues, but
these changes get some tests passing. The issue we encounter is that a
top() of a linear allocation area might be taken as a constant pointer
in CSA and then manipulated with either CSA code bumping the pointer or
a runtime call doing so. Since neither of these were taking pointer
alignment into account and enforcing the invariant that the top of the
area must always be pointer-aligned, it was tripping up an assertion
later on in the execution.

This commit re-uses the ALIGN_TO_ALLOCATION_ALIGNMENT() macro to always
align allocations to pointer size in the runtime, and generates code to
correctly align the top address in CSA for non-runtime calls. A known
issue with this code is that the original code will attempt to store a
map of a filler object in the alignment gap, but we cannot do this that
easily on a CHERI system. A comment is added to explain why with a FIXME
and TODO.