[execution] Enable deriving pages from sentries.
When unwinding the stack, we can encounter a situation where the
inner_pointer is in fact a sentry. This means we can't reliably use it
during the Contains() check for the code space, as it will cause a tag
fault. Instead, use a CHERI-specific check on CHERI systems.
Furthermore, if we have a sentry, we can't derive the page from the
sentry. Add a new Page::FromSentry() static method that derives a page
capability using a base capability from a linear allocation area and the
sentry as the address within that page.