[objects] Take into account the LSB bit in purecap ABI in Code::GetOffsetFromInstructionStart
When the PCC is sealed in purecap ABI, the LSB bit is set, leading
to an off-by-one error in the computation of the offset from the
instruction start to the current PCC address. This can cause errors
such as in OptimizedFrame::LookupExceptionHandlerInTable, where the
computed offset is used to look up an exception handler table.

(cherry picked from commit e04b2bb9a19574ff3ed1aa0d19434fbd6efeede3)