[toolchain] Run mksnapshot on a remote runner or use qemu usermode
This commit adds support for running mksnapshot on a remote host via
SSH, or via QEMU usermode. This is necessary to cross-compile purecap
V8 because of the bit-width difference of the build host and the target host.

For QEMU usermode, it is necessary to run with the --single-threaded flag
because currently QEMU CHERI doesn't support multithreading.

Co-authored-by: Domagoj Stolfa <ds815@cam.ac.uk>