[regexp] Fix how the link register is derived.
The link register that we pop is in fact an offset rather than a
capability. Instead of doing the complex capability re-deriving with
scvalue from the PCC, add the capability to the constant pool and encode
a ldr-literal + a capability add to derive the correct link register.

We might want to turn this into a sentry at some point, but the sensible
way to do that is currently unclear.